Move build_multitask_decoder to xm_transformer_unity.py
